migrate to tracker v2
