+++ core/tracker/tracker.go
@@ -6,8 +6,10 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"sort"
 
+	eth2http "github.com/attestantio/go-eth2-client/http"
+
+	"github.com/obolnetwork/charon/app/errors"
 	"github.com/obolnetwork/charon/app/log"
 	"github.com/obolnetwork/charon/app/z"
 	"github.com/obolnetwork/charon/core"
@@ -16,28 +18,33 @@ import (
 
 // Steps arranged in the order they are triggered in the core workflow.
 const (
-	zero              step = iota
-	scheduler              // Duty scheduled with definition
-	fetcher                // Duty data fetched
-	consensus              // Duty data consensus reached
-	validatorAPI           // Partial signed data from local VC submitted to vapi
-	parSigDBInternal       // Partial signed data from local VC stored in parsigdb
-	parSigEx               // Partial signed data from other VC received via parsigex
-	parSigDBThreshold      // Partial signed data threshold reached; emitted from parsigdb
-	sigAgg                 // Partial signed data aggregated; emitted from sigagg
+	zero             step = iota
+	fetcher               // Duty data fetched
+	consensus             // Duty data consensus reached
+	dutyDB                // Duty data stored in DutyDB
+	validatorAPI          // Partial signed data from local VC submitted to vapi
+	parSigDBInternal      // Partial signed data from local VC stored in parsigdb
+	parSigEx              // Partial signed data from local VC to/from other peers in cluster
+	parSigDBExternal      // Partial signed data from other VC stored in parsigdb
+	sigAgg                // Partial signed data aggregated; emitted from sigagg
+	aggSigDB              // Aggregated signed data stored in aggsigdb
+	bcast                 // Aggregated data submitted to beacon node
 	sentinel
 )
 
 var stepLabels = map[step]string{
-	zero:              "unknown",
-	scheduler:         "scheduler",
-	fetcher:           "fetcher",
-	consensus:         "consensus",
-	validatorAPI:      "validator_api",
-	parSigDBInternal:  "parsig_db_local",
-	parSigEx:          "parsig_exchange",
-	parSigDBThreshold: "parsig_db_threshold",
-	sigAgg:            "sig_aggregation",
+	zero:             "unknown",
+	fetcher:          "fetcher",
+	consensus:        "consensus",
+	dutyDB:           "duty_db",
+	validatorAPI:     "validator_api",
+	parSigDBInternal: "parsig_db_local",
+	parSigEx:         "parsig_ex",
+	parSigDBExternal: "parsig_db_external",
+	sigAgg:           "sig_aggregation",
+	aggSigDB:         "aggsig_db",
+	bcast:            "bcast",
+	sentinel:         "sentinel",
 }
 
 // step in the core workflow.
@@ -49,10 +56,10 @@ func (s step) String() string {
 
 // These constants are used for improving messages for why a duty failed.
 const (
-	// msgFetcher indicates a duty failed in the fetcher step when it failed
+	// msgFetcherBN indicates a duty failed in the fetcher step when it failed
 	// to fetch the required data from the beacon node API. This indicates a problem with
 	// the upstream beacon node.
-	msgFetcher = "couldn't fetch duty data from the beacon node"
+	msgFetcherBN = "couldn't fetch duty data from the beacon node"
 
 	// msgFetcherAggregatorNoAttData indicates an attestation aggregation duty failed in
 	// the fetcher step since it couldn't fetch the prerequisite attestation data. This
@@ -76,6 +83,11 @@ const (
 	// This indicates the associated prepare aggregation duty failed.
 	msgFetcherAggregatorFailedPrepare = "couldn't aggregate attestation due to failed prepare aggregator duty"
 
+	// msgFetcherAggregatorNoExternalPrepares indicates an attestation aggregation duty failed in
+	// the fetcher step since it couldn't fetch the prerequisite aggregated beacon committee selections.
+	// This indicates the associated prepare aggregation duty failed due to no partial beacon committee selections received from peers.
+	msgFetcherAggregatorNoExternalPrepares = "couldn't aggregate attestation due to no partial beacon committee selections received from peers"
+
 	// msgFetcherProposerFewRandaos indicates a block proposer duty failed in
 	// the fetcher step since it couldn't fetch the prerequisite aggregated RANDAO.
 	// This indicates the associated randao duty failed due to insufficient partial randao signatures
@@ -92,6 +104,11 @@ const (
 	// the fetcher step since it couldn't fetch the prerequisite aggregated RANDAO.
 	// This indicates the associated randao duty failed.
 	msgFetcherProposerFailedRandao = "couldn't propose block due to failed randao duty"
+
+	// msgFetcherProposerNoExternalRandaos indicates a block proposer duty failed in
+	// the fetcher step since it couldn't fetch the prerequisite aggregated RANDAO.
+	// This indicates the associated randao duty failed due to no partial randao signatures received from peers.
+	msgFetcherProposerNoExternalRandaos = "couldn't propose block due to no partial randao signatures received from peers"
 
 	// msgFetcherSyncContributionNoSyncMsg indicates a sync contribution duty failed in
 	// the fetcher step since it couldn't fetch the prerequisite sync message. This
@@ -115,10 +132,22 @@ const (
 	// This indicates the associated prepare sync contribution duty failed.
 	msgFetcherSyncContributionFailedPrepare = "couldn't fetch sync contribution due to failed prepare sync contribution duty"
 
+	// msgFetcherSyncContributionNoExternalPrepares indicates a sync contribution duty failed in
+	// the fetcher step since it couldn't fetch the prerequisite aggregated sync contribution selections.
+	// This indicates the associated prepare sync contribution duty failed due to no partial sync contribution selections received from peers.
+	msgFetcherSyncContributionNoExternalPrepares = "couldn't fetch sync contribution due to no partial sync contribution selections received from peers"
+
+	// msgFetcherError indicates duty failed in fetcher step with some unexpected error.
+	// This indicates a problem in charon as it is unexpected.
+	msgFetcherError = "couldn't fetch due to unexpected error"
+
 	// msgConsensus indicates a duty failed in consensus step.
 	// This could indicate that insufficient honest peers participated in consensus or p2p network
 	// connection problems.
 	msgConsensus = "consensus algorithm didn't complete"
+
+	// msgDutyDB indicates a bug in the DutyDB database as it is unexpected.
+	msgDutyDB = "bug: failed to store duty data in DutyDB"
 
 	// msgValidatorAPI indicates that partial signature we never submitted by the local
 	// validator client. This could indicate that the local validator client is offline,
@@ -128,11 +157,11 @@ const (
 
 	// msgParSigDBInternal indicates a bug in the partial signature database as it is unexpected.
 	// Note this may happen due to expiry race.
-	msgParSigDBInternal = "partial signature database didn't trigger partial signature exchange"
+	msgParSigDBInternal = "partial signature database didn't trigger partial signature exchange, this is unexpected"
 
-	// msgParSigEx indicates that no partial signature for the duty was received from any peer.
+	// msgParSigExReceive indicates that no partial signature for the duty was received from any peer.
 	// This indicates all peers are offline or p2p network connection problems.
-	msgParSigEx = "no partial signatures received from peers"
+	msgParSigExReceive = "no partial signatures received from peers"
 
 	// msgParSigDBInsufficient indicates that insufficient partial signatures for the duty was received from peers.
 	// This indicates problems with peers or p2p network connection problems.
@@ -146,9 +175,18 @@ const (
 	// This indicates a bug in charon as it is unexpected (for non-sync-committee-duties).
 	msgParSigDBInconsistent = "bug: inconsistent partial signatures received"
 
+	// msgParSigDBExternal indicates a bug in the partial signature database as it is unexpected.
+	msgParSigDBExternal = "bug: failed to store external partial signatures in parsigdb"
+
 	// msgSigAgg indicates that BLS threshold aggregation of sufficient partial signatures failed. This
 	// indicates inconsistent signed data. This indicates a bug in charon as it is unexpected.
 	msgSigAgg = "bug: threshold aggregation of partial signatures failed due to inconsistent signed data"
+
+	// msgAggSigDB indicates a bug in the aggregated signature database as it is unexpected.
+	msgAggSigDB = "bug: failed to store aggregated signature in aggsigdb"
+
+	// msgBcast indicates that beacon node returned an error while submitting aggregated duty signature to beacon node.
+	msgBcast = "failed to broadcast duty to beacon node"
 )
 
 // parsigsByMsg contains partial signatures grouped by message root grouped by pubkey.
@@ -156,8 +194,8 @@ type parsigsByMsg map[core.PubKey]map[[32]byte][]core.ParSignedData
 
 // MsgRootsConsistent returns true if the all partial signatures have consistent message roots.
 func (m parsigsByMsg) MsgRootsConsistent() bool {
-	for _, inner := range m {
-		if len(inner) > 1 {
+	for _, roots := range m {
+		if len(roots) > 1 {
 			return false
 		}
 	}
@@ -167,11 +205,12 @@ func (m parsigsByMsg) MsgRootsConsistent() bool {
 
 // event represents an event emitted by a core workflow step.
 type event struct {
-	duty   core.Duty
-	step   step
-	pubkey core.PubKey
+	duty    core.Duty
+	step    step
+	pubkey  core.PubKey
+	stepErr error
 
-	// parSig is an optional field only set by validatorAPI, parSigDBInternal and parSigEx events.
+	// parSig is an optional field only set by validatorAPI, parSigDBInternal and parSigExReceive events.
 	parSig *core.ParSignedData
 }
 
@@ -194,7 +233,7 @@ type Tracker struct {
 	parSigReporter func(ctx context.Context, duty core.Duty, parsigMsgs parsigsByMsg)
 
 	// failedDutyReporter instruments duty failures.
-	failedDutyReporter func(ctx context.Context, duty core.Duty, failed bool, step step, reason string)
+	failedDutyReporter func(ctx context.Context, duty core.Duty, failed bool, step step, reason string, err error)
 
 	// participationReporter instruments duty peer participation.
 	participationReporter func(ctx context.Context, duty core.Duty, failed bool, participatedShares map[int]bool, unexpectedPeers map[int]bool)
@@ -239,12 +278,12 @@ func (t *Tracker) Run(ctx context.Context) error {
 		case duty := <-t.analyser.C():
 			ctx := log.WithCtx(ctx, z.Any("duty", duty))
 
-			parsigs := analyseParSigs(ctx, t.events[duty])
+			parsigs := extractParSigs(ctx, t.events[duty])
 			t.parSigReporter(ctx, duty, parsigs)
 
 			// Analyse failed duties
-			failed, failedStep, failedMsg := analyseDutyFailed(duty, t.events, parsigs.MsgRootsConsistent())
-			t.failedDutyReporter(ctx, duty, failed, failedStep, failedMsg)
+			failed, failedStep, failedMsg, failedErr := analyseDutyFailed(duty, t.events, parsigs.MsgRootsConsistent())
+			t.failedDutyReporter(ctx, duty, failed, failedStep, failedMsg, failedErr)
 
 			// Analyse peer participation
 			participatedShares, unexpectedShares := analyseParticipation(duty, t.events)
@@ -255,90 +294,252 @@ func (t *Tracker) Run(ctx context.Context) error {
 	}
 }
 
-// dutyFailedStep returns true if the duty failed. It also returns the step where the duty got stuck.
-// If the duty didn't fail, it returns false and the zero step.
+// dutyFailedStep returns true if the duty failed. It also returns the step where the
+// duty got stuck and the last error that component returned.
+// If the duty didn't fail, it returns false and the zero step and a nil error.
 // It assumes that all the input events are for a single duty.
 // If the input events slice is empty, it returns true and the zero step.
-func dutyFailedStep(es []event) (bool, step) {
+func dutyFailedStep(es []event) (bool, step, error) {
 	if len(es) == 0 {
-		return true, zero // Duty failed since no events.
+		return true, zero, nil // Duty failed since no events.
 	}
 
-	// Copy and sort in reverse order (see step order above).
-	clone := append([]event(nil), es...)
-	sort.Slice(clone, func(i, j int) bool {
-		return clone[i].step > clone[j].step
-	})
-
-	step := clone[0].step
-	if step == sigAgg {
-		return false, zero
+	// Group events by step.
+	eventsByStep := make(map[step][]event)
+	for _, e := range es {
+		eventsByStep[e.step] = append(eventsByStep[e.step], e)
 	}
 
-	return true, step + 1
+	// Find last failed/successful step.
+	var lastEvent event
+	for step := sentinel - 1; step > zero; step-- {
+		if len(eventsByStep[step]) == 0 {
+			continue
+		}
+
+		lastEvent = eventsByStep[step][len(eventsByStep[step])-1]
+
+		break
+	}
+
+	// Final step was successful.
+	if lastEvent.step == bcast && lastEvent.stepErr == nil {
+		return false, zero, nil
+	}
+
+	return true, lastEvent.step, lastEvent.stepErr
 }
 
 // analyseDutyFailed detects if the given duty failed.
 //
 // It returns true if the duty failed as well as the step
-// where the duty got stuck and a human friendly error message explaining why.
+// where the duty got stuck and a human friendly error message explaining why
+// as well as the error reported by the step/component.
 //
 // It returns false if the duty didn't fail, i.e., the duty
-// didn't get stuck and completed the sigAgg step.
-func analyseDutyFailed(duty core.Duty, allEvents map[core.Duty][]event, msgRootConsistent bool) (bool, step, string) {
-	failed, step := dutyFailedStep(allEvents[duty])
+// didn't get stuck and completed the bcast step.
+func analyseDutyFailed(duty core.Duty, allEvents map[core.Duty][]event, msgRootConsistent bool,
+) (bool, step, string, error) {
+	failed, failedStep, failedErr := dutyFailedStep(allEvents[duty])
 	if !failed {
-		return false, zero, ""
+		return false, failedStep, "", nil
 	}
 
-	var msg string
-	switch step {
+	failedMsg := fmt.Sprintf("unexpected failure for %s duty at %s step", duty, failedStep)
+	switch failedStep {
 	case fetcher:
-		return analyseFetcherFailed(duty, allEvents)
+		return analyseFetcherFailed(duty, allEvents, failedErr)
 	case consensus:
-		msg = msgConsensus
-	case validatorAPI:
-		msg = msgValidatorAPI
-	case parSigDBInternal:
-		msg = msgParSigDBInternal
-	case parSigEx:
-		msg = msgParSigEx
-	case parSigDBThreshold:
-		if msgRootConsistent {
-			msg = msgParSigDBInsufficient
+		if failedErr != nil {
+			failedMsg = msgConsensus
+		}
+	case dutyDB:
+		if failedErr != nil {
+			failedMsg = msgDutyDB
 		} else {
-			msg = msgParSigDBInconsistent
-			if expectInconsistentParSigs(duty.Type) {
-				msg = msgParSigDBInconsistentSync
-			}
+			failedStep = validatorAPI
+			failedMsg = msgValidatorAPI
+		}
+	case parSigDBInternal:
+		failedMsg = msgParSigDBInternal
+	case parSigEx:
+		if failedErr == nil {
+			failedMsg = msgParSigExReceive
+		}
+	case parSigDBExternal:
+		if failedErr != nil {
+			return true, parSigDBExternal, msgParSigDBExternal, failedErr
 		}
 
-		return true, parSigDBThreshold, msg
+		if msgRootConsistent {
+			failedMsg = msgParSigDBInsufficient
+		} else {
+			failedMsg = msgParSigDBInconsistent
+			if expectInconsistentParSigs(duty.Type) {
+				failedMsg = msgParSigDBInconsistentSync
+			}
+		}
 	case sigAgg:
-		msg = msgSigAgg
+		if failedErr != nil {
+			failedMsg = msgSigAgg
+		}
+	case aggSigDB:
+		failedMsg = msgAggSigDB
+	case bcast:
+		failedMsg = msgBcast
 	case zero:
-		msg = fmt.Sprintf("no events for %s duty", duty.String()) // This should never happen.
+		failedMsg = fmt.Sprintf("no events for %s duty", duty.String())
 	default:
-		msg = fmt.Sprintf("%s duty failed at %s", duty.String(), step.String())
+		failedMsg = fmt.Sprintf("%s duty failed at %s", duty.String(), failedStep.String()) // This should never happen.
 	}
 
-	return true, step, msg
+	return true, failedStep, failedMsg, failedErr
 }
 
-// analyseParSigs returns a mapping of partial signed data messages by peers (share index) by validator (pubkey).
-func analyseParSigs(ctx context.Context, events []event) parsigsByMsg {
+// analyseFetcherFailed returns whether the duty that got stuck in fetcher actually failed
+// and the reason which might actually be due a pre-requisite duty that failed.
+func analyseFetcherFailed(duty core.Duty, allEvents map[core.Duty][]event, fetchErr error) (bool, step, string, error) {
+	failedMsg := msgFetcherError
+	// Check for beacon api errors.
+	var eth2Error eth2http.Error
+	if errors.As(fetchErr, &eth2Error) {
+		failedMsg = msgFetcherBN
+	} else if !errors.Is(fetchErr, context.Canceled) && !errors.Is(fetchErr, context.DeadlineExceeded) {
+		failedMsg = msgFetcherError
+	}
+
+	// Proposer duties depend on randao duty, so check if that was why it failed.
+	if duty.Type == core.DutyProposer || duty.Type == core.DutyBuilderProposer {
+		return analyseFetcherFailedProposer(duty, allEvents, fetchErr)
+	}
+
+	// Duty aggregator depend on prepare aggregator duty, so check if that was why it failed.
+	if duty.Type == core.DutyAggregator {
+		return analyseFetcherFailedAggregator(duty, allEvents, fetchErr)
+	}
+
+	// Duty sync contribution depends on prepare sync contribution duty, so check if that was why it failed.
+	if duty.Type == core.DutySyncContribution {
+		return analyseFetcherFailedSyncContribution(duty, allEvents, fetchErr)
+	}
+
+	return true, fetcher, failedMsg, fetchErr
+}
+
+// analyseFetcherFailed returns the reason behind why proposer duty failed which might actually
+// be due to randao duty failed.
+func analyseFetcherFailedProposer(duty core.Duty, allEvents map[core.Duty][]event, fetchErr error) (bool, step, string, error) {
+	failedMsg := msgFetcherError
+
+	// Proposer duties will fail if core.DutyRandao fails.
+	// Ignoring error as it will be handled in DutyRandao analysis.
+	randaoFailed, randaoStep, _ := dutyFailedStep(allEvents[core.NewRandaoDuty(duty.Slot)])
+	if randaoFailed {
+		switch randaoStep {
+		case parSigEx:
+			failedMsg = msgFetcherProposerNoExternalRandaos
+		case parSigDBExternal:
+			failedMsg = msgFetcherProposerFewRandaos
+		case zero:
+			failedMsg = msgFetcherProposerZeroRandaos
+		default:
+			failedMsg = msgFetcherProposerFailedRandao
+		}
+	}
+
+	return true, fetcher, failedMsg, fetchErr
+}
+
+// analyseFetcherFailedAggregator returns the reason behind why aggregator duty failed which might actually
+// be due to prepare aggregator duty or attester duty failed.
+func analyseFetcherFailedAggregator(duty core.Duty, allEvents map[core.Duty][]event, fetchErr error) (bool, step, string, error) {
+	failedMsg := msgFetcherError
+
+	// No aggregators present for this slot.
+	if fetchErr == nil {
+		return false, fetcher, "", nil
+	}
+
+	// Aggregator duties will fail if core.DutyPrepareAggregator fails.
+	// Ignoring error as it will be handled in DutyPrepareAggregator duty analysis.
+	prepAggFailed, prepAggStep, _ := dutyFailedStep(allEvents[core.NewPrepareAggregatorDuty(duty.Slot)])
+	if prepAggFailed {
+		switch prepAggStep {
+		case parSigEx:
+			failedMsg = msgFetcherAggregatorNoExternalPrepares
+		case parSigDBExternal:
+			failedMsg = msgFetcherAggregatorFewPrepares
+		case zero:
+			failedMsg = msgFetcherAggregatorZeroPrepares
+		default:
+			failedMsg = msgFetcherAggregatorFailedPrepare
+		}
+
+		return true, fetcher, failedMsg, fetchErr
+	}
+
+	// Aggregator duties will fail if no attestation data in DutyDB.
+	// Ignoring error as it will be handled in DutyAttester analysis.
+	attFailed, attStep, _ := dutyFailedStep(allEvents[core.NewAttesterDuty(duty.Slot)])
+	if attFailed && attStep <= dutyDB {
+		failedMsg = msgFetcherAggregatorNoAttData
+	}
+
+	return true, fetcher, failedMsg, fetchErr
+}
+
+// analyseFetcherFailedSyncContribution returns the reason behind why sync contribution duty failed which might actually
+// be due to prepare sync contribution duty or sync message duty failed.
+func analyseFetcherFailedSyncContribution(duty core.Duty, allEvents map[core.Duty][]event, fetchErr error) (bool, step, string, error) {
+	failedMsg := msgFetcherError
+
+	// No sync committee aggregators present for this slot.
+	if fetchErr == nil {
+		return false, fetcher, "", nil
+	}
+
+	// Sync contribution duties will fail if core.DutyPrepareSyncContribution fails.
+	// Ignoring error as it will be handled in DutyPrepareSyncContribution analysis.
+	prepSyncConFailed, prepSyncConStep, _ := dutyFailedStep(allEvents[core.NewPrepareSyncContributionDuty(duty.Slot)])
+	if prepSyncConFailed {
+		switch prepSyncConStep {
+		case parSigEx:
+			failedMsg = msgFetcherSyncContributionNoExternalPrepares
+		case parSigDBExternal:
+			failedMsg = msgFetcherSyncContributionFewPrepares
+		case zero:
+			failedMsg = msgFetcherSyncContributionZeroPrepares
+		default:
+			failedMsg = msgFetcherSyncContributionFailedPrepare
+		}
+
+		return true, fetcher, failedMsg, fetchErr
+	}
+
+	// Sync contribution duties will fail if no sync message in AggSigDB.
+	// Ignoring error as it will be handled in DutySyncMessage analysis.
+	syncMsgFailed, syncMsgStep, _ := dutyFailedStep(allEvents[core.NewSyncMessageDuty(duty.Slot)])
+	if syncMsgFailed && syncMsgStep <= aggSigDB {
+		failedMsg = msgFetcherSyncContributionNoSyncMsg
+	}
+
+	return true, fetcher, failedMsg, fetchErr
+}
+
+// extractParSigs returns a mapping of unique partial signed data messages by peers (share index) by validator (pubkey).
+func extractParSigs(ctx context.Context, events []event) parsigsByMsg {
 	type dedupKey struct {
 		Pubkey   core.PubKey
 		ShareIdx int
 	}
 	var (
 		dedup = make(map[dedupKey]bool)
-		datas = make(map[core.PubKey]map[[32]byte][]core.ParSignedData)
+		resp  = make(parsigsByMsg)
 	)
 
 	for _, e := range events {
 		if e.parSig == nil {
-			continue
+			continue // Ignore events without parsigs.
 		}
 		key := dedupKey{Pubkey: e.pubkey, ShareIdx: e.parSig.ShareIdx}
 		if dedup[key] {
@@ -352,112 +553,22 @@ func analyseParSigs(ctx context.Context, events []event) parsigsByMsg {
 			continue // Just log and ignore as this is highly unlikely and non-critical code.
 		}
 
-		inner, ok := datas[e.pubkey]
+		inner, ok := resp[e.pubkey]
 		if !ok {
 			inner = make(map[[32]byte][]core.ParSignedData)
 		}
 		inner[root] = append(inner[root], *e.parSig)
-		datas[e.pubkey] = inner
+		resp[e.pubkey] = inner
 	}
 
-	return datas
-}
-
-// analyseFetcherFailed returns whether the duty that got stuck in fetcher actually failed
-// and the reason which might actually be due a pre-requisite duty that failed.
-func analyseFetcherFailed(duty core.Duty, allEvents map[core.Duty][]event) (bool, step, string) {
-	msg := msgFetcher
-
-	// Proposer duties depend on randao duty, so check if that was why it failed.
-	if duty.Type == core.DutyProposer || duty.Type == core.DutyBuilderProposer {
-		// Proposer duties will fail if core.DutyRandao fails
-		randaoFailed, randaoStep := dutyFailedStep(allEvents[core.NewRandaoDuty(duty.Slot)])
-		if randaoFailed {
-			switch randaoStep {
-			case parSigDBThreshold:
-				msg = msgFetcherProposerFewRandaos
-			case zero:
-				msg = msgFetcherProposerZeroRandaos
-			default:
-				msg = msgFetcherProposerFailedRandao
-			}
-		}
-
-		return true, fetcher, msg
-	}
-
-	// Duty aggregator depend on prepare aggregator duty, so check if that was why it failed.
-	if duty.Type == core.DutyAggregator {
-		// Aggregator duties will fail if core.DutyPrapareAggregator fails
-		prepAggFailed, prepAggStep := dutyFailedStep(allEvents[core.NewPrepareAggregatorDuty(duty.Slot)])
-		if prepAggFailed {
-			switch prepAggStep {
-			case parSigDBThreshold:
-				msg = msgFetcherAggregatorFewPrepares
-			case zero:
-				msg = msgFetcherAggregatorZeroPrepares
-			default:
-				msg = msgFetcherAggregatorFailedPrepare
-			}
-
-			return true, fetcher, msg
-		}
-
-		// Aggregator duties will fail if no attestation data in DutyDB
-		attFailed, attStep := dutyFailedStep(allEvents[core.NewAttesterDuty(duty.Slot)])
-		if attFailed && attStep <= consensus {
-			// Note we do not handle the edge case of the local peer failing to store attestation data
-			// but the attester duty succeeding in any case due to external peer partial signatures.
-			return true, fetcher, msgFetcherAggregatorNoAttData
-		}
-
-		// TODO(corver): We cannot distinguish between "no aggregators for slot"
-		//  and "failed fetching aggregated attestation from BN".
-		//
-		// Assume no aggregators for slot as this is very common.
-		return false, fetcher, ""
-	}
-
-	// Duty sync contribution depends on prepare sync contribution duty, so check if that was why it failed.
-	if duty.Type == core.DutySyncContribution {
-		// Sync contribution duties will fail if core.DutyPrepareSyncContribution fails.
-		prepSyncConFailed, prepSyncConStep := dutyFailedStep(allEvents[core.NewPrepareSyncContributionDuty(duty.Slot)])
-		if prepSyncConFailed {
-			switch prepSyncConStep {
-			case parSigDBThreshold:
-				msg = msgFetcherSyncContributionFewPrepares
-			case zero:
-				msg = msgFetcherSyncContributionZeroPrepares
-			default:
-				msg = msgFetcherSyncContributionFailedPrepare
-			}
-
-			return true, fetcher, msg
-		}
-
-		// Sync contribution duties will fail if no sync message in AggSigDB.
-		syncMsgFailed, syncMsgStep := dutyFailedStep(allEvents[core.NewSyncMessageDuty(duty.Slot)])
-		if syncMsgFailed && syncMsgStep <= sigAgg {
-			// Note we do not handle the edge case of the local peer failing to store sync message
-			// but the sync message duty succeeding in any case due to external peer partial signatures.
-			return true, fetcher, msgFetcherSyncContributionNoSyncMsg
-		}
-
-		// TODO(dhruv): We cannot distinguish between "no sync committee aggregators for slot"
-		//  and "failed fetching sync committee contribution from BN".
-		//
-		// Assume no aggregators for slot as this is very common.
-		return false, fetcher, ""
-	}
-
-	return true, fetcher, msg
+	return resp
 }
 
 // newFailedDutyReporter returns failed duty reporter which instruments failed duties.
-func newFailedDutyReporter() func(ctx context.Context, duty core.Duty, failed bool, step step, reason string) {
+func newFailedDutyReporter() func(ctx context.Context, duty core.Duty, failed bool, step step, reason string, err error) {
 	var loggedNoSelections bool
 
-	return func(ctx context.Context, duty core.Duty, failed bool, step step, reason string) {
+	return func(ctx context.Context, duty core.Duty, failed bool, step step, reason string, err error) {
 		counter := failedCounter.WithLabelValues(duty.Type.String())
 		counter.Add(0) // Zero the metric so first failure shows in grafana.
 
@@ -483,9 +594,10 @@ func newFailedDutyReporter() func(ctx context.Context, duty core.Duty, failed bo
 			return
 		}
 
-		log.Warn(ctx, "Duty failed", nil,
+		log.Warn(ctx, "Duty failed", err,
 			z.Any("step", step),
-			z.Str("reason", reason))
+			z.Str("reason", reason),
+		)
 
 		counter.Inc()
 	}
@@ -499,8 +611,8 @@ func analyseParticipation(duty core.Duty, allEvents map[core.Duty][]event) (map[
 
 	for _, e := range allEvents[duty] {
 		// If we get a parSigDBInternal event, then the current node participated successfully.
-		// If we get a parSigEx event, then the corresponding peer with e.shareIdx participated successfully.
-		if e.step == parSigEx || e.step == parSigDBInternal {
+		// If we get a parSigDBExternal event, then the corresponding peer with e.shareIdx participated successfully.
+		if e.step == parSigDBExternal || e.step == parSigDBInternal {
 			if !isParSigEventExpected(duty, e.pubkey, allEvents) {
 				unexpectedShares[e.parSig.ShareIdx] = true
 				continue
@@ -524,7 +636,7 @@ func isParSigEventExpected(duty core.Duty, pubkey core.PubKey, allEvents map[cor
 	// scheduled returns true if the provided duty type was scheduled for the above slot and pubkey.
 	scheduled := func(typ core.DutyType) bool {
 		for _, e := range allEvents[core.Duty{Slot: duty.Slot, Type: typ}] {
-			if e.step == scheduler && e.pubkey == pubkey {
+			if e.step == fetcher && e.pubkey == pubkey {
 				return true
 			}
 		}
@@ -591,182 +703,148 @@ func newParticipationReporter(peers []p2p.Peer) func(context.Context, core.Duty,
 	}
 }
 
-// SchedulerEvent inputs event from core.Scheduler step.
-func (t *Tracker) SchedulerEvent(ctx context.Context, duty core.Duty, defSet core.DutyDefinitionSet) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
-	for pubkey := range defSet {
+// FetcherFetched implements core.Tracker interface.
+func (t *Tracker) FetcherFetched(duty core.Duty, set core.DutyDefinitionSet, stepErr error) {
+	for pubkey := range set {
 		select {
-		case <-ctx.Done():
-			return ctx.Err()
 		case <-t.quit:
-			return nil
+			return
 		case t.input <- event{
-			duty:   duty,
-			step:   scheduler,
-			pubkey: pubkey,
+			duty:    duty,
+			step:    fetcher,
+			pubkey:  pubkey,
+			stepErr: stepErr,
 		}:
 		}
 	}
-
-	return nil
 }
 
-// FetcherEvent inputs event from core.Fetcher step.
-func (t *Tracker) FetcherEvent(ctx context.Context, duty core.Duty, data core.UnsignedDataSet) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
-	for pubkey := range data {
+// ConsensusProposed implements core.Tracker interface.
+func (t *Tracker) ConsensusProposed(duty core.Duty, set core.UnsignedDataSet, stepErr error) {
+	for pubkey := range set {
 		select {
-		case <-ctx.Done():
-			return ctx.Err()
 		case <-t.quit:
-			return nil
+			return
 		case t.input <- event{
-			duty:   duty,
-			step:   fetcher,
-			pubkey: pubkey,
+			duty:    duty,
+			step:    consensus,
+			pubkey:  pubkey,
+			stepErr: stepErr,
 		}:
 		}
 	}
-
-	return nil
 }
 
-// ConsensusEvent inputs event from core.Consensus step.
-func (t *Tracker) ConsensusEvent(ctx context.Context, duty core.Duty, data core.UnsignedDataSet) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
-	for pubkey := range data {
+// DutyDBStored implements core.Tracker interface.
+func (t *Tracker) DutyDBStored(duty core.Duty, set core.UnsignedDataSet, stepErr error) {
+	for pubkey := range set {
 		select {
-		case <-ctx.Done():
-			return ctx.Err()
 		case <-t.quit:
-			return nil
+			return
 		case t.input <- event{
-			duty:   duty,
-			step:   consensus,
-			pubkey: pubkey,
+			duty:    duty,
+			step:    dutyDB,
+			pubkey:  pubkey,
+			stepErr: stepErr,
 		}:
 		}
 	}
-
-	return nil
 }
 
-// ValidatorAPIEvent inputs events from core.ValidatorAPI step.
-func (t *Tracker) ValidatorAPIEvent(ctx context.Context, duty core.Duty, data core.ParSignedDataSet) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
-	for pubkey, parSig := range data {
-		parSig := parSig // Copy loop iteration values
+// ParSigDBStoredInternal implements core.Tracker interface.
+func (t *Tracker) ParSigDBStoredInternal(duty core.Duty, set core.ParSignedDataSet, stepErr error) {
+	for pubkey, parSig := range set {
+		parSig := parSig
 		select {
-		case <-ctx.Done():
-			return ctx.Err()
 		case <-t.quit:
-			return nil
+			return
 		case t.input <- event{
-			duty:   duty,
-			step:   validatorAPI,
-			pubkey: pubkey,
-			parSig: &parSig,
+			duty:    duty,
+			step:    parSigDBInternal,
+			pubkey:  pubkey,
+			parSig:  &parSig,
+			stepErr: stepErr,
 		}:
 		}
 	}
-
-	return nil
 }
 
-// ParSigExEvent inputs event from core.ParSigEx step event for other VC submitted parsigs.
-func (t *Tracker) ParSigExEvent(ctx context.Context, duty core.Duty, data core.ParSignedDataSet) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
-	for pubkey, parSig := range data {
-		parSig := parSig // Copy loop iteration values
+// ParSigExBroadcasted implements core.Tracker interface.
+func (t *Tracker) ParSigExBroadcasted(duty core.Duty, set core.ParSignedDataSet, stepErr error) {
+	for pubkey, parSig := range set {
+		parSig := parSig
 		select {
-		case <-ctx.Done():
-			return ctx.Err()
 		case <-t.quit:
-			return nil
+			return
 		case t.input <- event{
-			duty:   duty,
-			step:   parSigEx,
-			pubkey: pubkey,
-			parSig: &parSig,
+			duty:    duty,
+			step:    parSigEx,
+			pubkey:  pubkey,
+			parSig:  &parSig,
+			stepErr: stepErr,
 		}:
 		}
 	}
-
-	return nil
 }
 
-// ParSigDBInternalEvent inputs events from core.ParSigDB step event for local VC submitted parsigs.
-func (t *Tracker) ParSigDBInternalEvent(ctx context.Context, duty core.Duty, data core.ParSignedDataSet) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
-	for pubkey, parSig := range data {
-		parSig := parSig // Copy loop iteration values
+// ParSigDBStoredExternal implements core.Tracker interface.
+func (t *Tracker) ParSigDBStoredExternal(duty core.Duty, set core.ParSignedDataSet, stepErr error) {
+	for pubkey, parSig := range set {
+		parSig := parSig
 		select {
-		case <-ctx.Done():
-			return ctx.Err()
 		case <-t.quit:
-			return nil
+			return
 		case t.input <- event{
-			duty:   duty,
-			step:   parSigDBInternal,
-			pubkey: pubkey,
-			parSig: &parSig,
+			duty:    duty,
+			step:    parSigDBExternal,
+			pubkey:  pubkey,
+			parSig:  &parSig,
+			stepErr: stepErr,
 		}:
 		}
 	}
-
-	return nil
 }
 
-// ParSigDBThresholdEvent inputs event from core.ParSigDB step for threshold emitted parsigs.
-func (t *Tracker) ParSigDBThresholdEvent(ctx context.Context, duty core.Duty, pubkey core.PubKey, _ []core.ParSignedData) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
+// SigAggAggregated implements core.Tracker interface.
+func (t *Tracker) SigAggAggregated(duty core.Duty, pubkey core.PubKey, _ []core.ParSignedData, stepErr error) {
 	select {
-	case <-ctx.Done():
-		return ctx.Err()
 	case <-t.quit:
-		return nil
+		return
 	case t.input <- event{
-		duty:   duty,
-		step:   parSigDBThreshold,
-		pubkey: pubkey,
+		duty:    duty,
+		step:    sigAgg,
+		pubkey:  pubkey,
+		stepErr: stepErr,
 	}:
 	}
-
-	return nil
 }
 
-// SigAggEvent inputs event from core.SigAgg step.
-func (t *Tracker) SigAggEvent(ctx context.Context, duty core.Duty, pubkey core.PubKey, _ core.SignedData) error {
-	if ctx.Err() != nil {
-		return nil //nolint:nilerr // Ignore event if expired.
-	}
+// AggSigSBStored implements core.Tracker interface.
+func (t *Tracker) AggSigDBStored(duty core.Duty, pubkey core.PubKey, _ core.SignedData, stepErr error) {
 	select {
-	case <-ctx.Done():
-		return ctx.Err()
 	case <-t.quit:
-		return nil
+		return
 	case t.input <- event{
-		duty:   duty,
-		step:   sigAgg,
-		pubkey: pubkey,
+		duty:    duty,
+		step:    aggSigDB,
+		pubkey:  pubkey,
+		stepErr: stepErr,
 	}:
 	}
+}
 
-	return nil
+// BroadcasterBroadcast implements core.Tracker interface.
+func (t *Tracker) BroadcasterBroadcast(duty core.Duty, pubkey core.PubKey, _ core.SignedData, stepErr error) {
+	select {
+	case <-t.quit:
+		return
+	case t.input <- event{
+		duty:    duty,
+		step:    bcast,
+		pubkey:  pubkey,
+		stepErr: stepErr,
+	}:
+	}
 }
 
 func reportParSigs(ctx context.Context, duty core.Duty, parsigMsgs parsigsByMsg) {
