+++ core/tracker/tracker_internal_test.go
@@ -4,13 +4,17 @@ package tracker
 
 import (
 	"context"
+	"fmt"
+	"net/http"
 	"reflect"
 	"testing"
 
 	eth2v1 "github.com/attestantio/go-eth2-client/api/v1"
+	eth2http "github.com/attestantio/go-eth2-client/http"
 	eth2p0 "github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/stretchr/testify/require"
 
+	"github.com/obolnetwork/charon/app/errors"
 	"github.com/obolnetwork/charon/core"
 	"github.com/obolnetwork/charon/p2p"
 	"github.com/obolnetwork/charon/testutil"
@@ -30,13 +34,15 @@ func TestTrackerFailedDuty(t *testing.T) {
 		ctx, cancel := context.WithCancel(context.Background())
 		analyser := testDeadliner{deadlineChan: make(chan core.Duty)}
 		deleter := testDeadliner{deadlineChan: make(chan core.Duty)}
+		consensusErr := errors.New("consensus error")
 
 		count := 0
-		failedDutyReporter := func(_ context.Context, failedDuty core.Duty, isFailed bool, step step, msg string) {
+		failedDutyReporter := func(_ context.Context, failedDuty core.Duty, isFailed bool, step step, msg string, err error) {
 			require.Equal(t, testData[0].duty, failedDuty)
 			require.True(t, isFailed)
 			require.Equal(t, consensus, step)
 			require.Equal(t, msg, msgConsensus)
+			require.ErrorIs(t, err, consensusErr)
 			count++
 
 			if count == len(testData) {
@@ -52,8 +58,8 @@ func TestTrackerFailedDuty(t *testing.T) {
 
 		go func() {
 			for _, td := range testData {
-				require.NoError(t, tr.SchedulerEvent(ctx, td.duty, td.defSet))
-				require.NoError(t, tr.FetcherEvent(ctx, td.duty, td.unsignedDataSet))
+				tr.FetcherFetched(td.duty, td.defSet, nil)
+				tr.ConsensusProposed(td.duty, td.unsignedDataSet, consensusErr)
 
 				// Explicitly mark the current duty as deadlined.
 				analyser.deadlineChan <- td.duty
@@ -72,7 +78,7 @@ func TestTrackerFailedDuty(t *testing.T) {
 		deleter := testDeadliner{deadlineChan: make(chan core.Duty)}
 
 		count := 0
-		failedDutyReporter := func(_ context.Context, failedDuty core.Duty, isFailed bool, step step, msg string) {
+		failedDutyReporter := func(_ context.Context, failedDuty core.Duty, isFailed bool, step step, msg string, _ error) {
 			require.Equal(t, testData[0].duty, failedDuty)
 			require.False(t, isFailed)
 			require.Equal(t, zero, step)
@@ -92,15 +98,14 @@ func TestTrackerFailedDuty(t *testing.T) {
 
 		go func() {
 			for _, td := range testData {
-				require.NoError(t, tr.SchedulerEvent(ctx, td.duty, td.defSet))
-				require.NoError(t, tr.FetcherEvent(ctx, td.duty, td.unsignedDataSet))
-				require.NoError(t, tr.ConsensusEvent(ctx, td.duty, td.unsignedDataSet))
-				require.NoError(t, tr.ValidatorAPIEvent(ctx, td.duty, td.parSignedDataSet))
-				require.NoError(t, tr.ParSigDBInternalEvent(ctx, td.duty, td.parSignedDataSet))
-				require.NoError(t, tr.ParSigExEvent(ctx, td.duty, td.parSignedDataSet))
+				tr.FetcherFetched(td.duty, td.defSet, nil)
+				tr.ConsensusProposed(td.duty, td.unsignedDataSet, nil)
+				tr.DutyDBStored(td.duty, td.unsignedDataSet, nil)
+				tr.ParSigDBStoredInternal(td.duty, td.parSignedDataSet, nil)
+				tr.ParSigDBStoredExternal(td.duty, td.parSignedDataSet, nil)
 				for _, pubkey := range pubkeys {
-					require.NoError(t, tr.ParSigDBThresholdEvent(ctx, td.duty, pubkey, nil))
-					require.NoError(t, tr.SigAggEvent(ctx, td.duty, pubkey, nil))
+					tr.SigAggAggregated(td.duty, pubkey, nil, nil)
+					tr.BroadcasterBroadcast(td.duty, pubkey, nil, nil)
 				}
 
 				// Explicitly mark the current duty as deadlined.
@@ -124,94 +129,124 @@ func TestAnalyseDutyFailed(t *testing.T) {
 
 	t.Run("Failed", func(t *testing.T) {
 		// Failed at fetcher
+		fetcherErr := errors.New("fetcher failed")
 		events := map[core.Duty][]event{
 			attDuty: {
 				{
-					duty: attDuty,
-					step: scheduler,
+					duty:    attDuty,
+					step:    fetcher,
+					stepErr: errors.New("fetcher failed"),
 				},
 			},
 		}
 
-		failed, step, msg := analyseDutyFailed(attDuty, events, true)
+		failed, step, msg, err := analyseDutyFailed(attDuty, events, true)
+		require.ErrorAs(t, err, &fetcherErr)
 		require.True(t, failed)
 		require.Equal(t, step, fetcher)
-		require.Equal(t, msg, msgFetcher)
+		require.Contains(t, msg, msgFetcherError)
 
 		// Failed at consensus
+		consensusErr := errors.New("consensus failed")
 		events[attDuty] = append(events[attDuty], event{
-			duty: attDuty,
-			step: fetcher,
+			duty:    attDuty,
+			step:    consensus,
+			stepErr: errors.New("consensus failed"),
 		})
 
-		failed, step, msg = analyseDutyFailed(attDuty, events, true)
+		failed, step, msg, err = analyseDutyFailed(attDuty, events, true)
+		require.ErrorAs(t, err, &consensusErr)
 		require.True(t, failed)
 		require.Equal(t, step, consensus)
-		require.Equal(t, msg, msgConsensus)
+		require.Contains(t, msg, msgConsensus)
 
 		// Failed at validatorAPI
 		events[attDuty] = append(events[attDuty], event{
 			duty: attDuty,
-			step: consensus,
+			step: dutyDB,
 		})
 
-		failed, step, msg = analyseDutyFailed(attDuty, events, true)
+		failed, step, msg, err = analyseDutyFailed(attDuty, events, true)
+		require.NoError(t, err)
 		require.True(t, failed)
 		require.Equal(t, step, validatorAPI)
 		require.Equal(t, msg, msgValidatorAPI)
 
 		// Failed at parsigDBInternal
+		parSigDBInternalErr := errors.New("parsigdb_internal failed")
 		events[attDuty] = append(events[attDuty], event{
-			duty: attDuty,
-			step: validatorAPI,
+			duty:    attDuty,
+			step:    parSigDBInternal,
+			stepErr: errors.New("parsigdb_internal failed"),
 		})
 
-		failed, step, msg = analyseDutyFailed(attDuty, events, true)
+		failed, step, msg, err = analyseDutyFailed(attDuty, events, true)
+		require.ErrorAs(t, err, &parSigDBInternalErr)
 		require.True(t, failed)
 		require.Equal(t, step, parSigDBInternal)
-		require.Equal(t, msg, msgParSigDBInternal)
+		require.Contains(t, msg, msgParSigDBInternal)
 
 		// Failed at parsigEx
-		events[attDuty] = append(events[attDuty], event{
-			duty: attDuty,
-			step: parSigDBInternal,
-		})
-
-		failed, step, msg = analyseDutyFailed(attDuty, events, true)
-		require.True(t, failed)
-		require.Equal(t, step, parSigEx)
-		require.Equal(t, msg, msgParSigEx)
-
-		// Failed at parsigDBThreshold
 		events[attDuty] = append(events[attDuty], event{
 			duty: attDuty,
 			step: parSigEx,
 		})
 
-		failed, step, msg = analyseDutyFailed(attDuty, events, true)
+		failed, step, msg, err = analyseDutyFailed(attDuty, events, true)
+		require.NoError(t, err)
 		require.True(t, failed)
-		require.Equal(t, step, parSigDBThreshold)
+		require.Equal(t, step, parSigEx)
+		require.Equal(t, msg, msgParSigExReceive)
+
+		// Failed at parsigDBInternal
+		parSigDBExternalErr := errors.New("parsigdb_external failed")
+		events[attDuty] = append(events[attDuty], event{
+			duty:    attDuty,
+			step:    parSigDBExternal,
+			stepErr: errors.New("parsigdb_external failed"),
+		})
+
+		failed, step, msg, err = analyseDutyFailed(attDuty, events, true)
+		require.ErrorAs(t, err, &parSigDBExternalErr)
+		require.True(t, failed)
+		require.Equal(t, step, parSigDBExternal)
+		require.Contains(t, msg, msgParSigDBExternal)
+
+		// Failed at parsigDBThreshold
+		events[attDuty] = append(events[attDuty], event{
+			duty: attDuty,
+			step: parSigDBExternal,
+		})
+
+		failed, step, msg, err = analyseDutyFailed(attDuty, events, true)
+		require.NoError(t, err)
+		require.True(t, failed)
+		require.Equal(t, step, parSigDBExternal)
 		require.Equal(t, msg, msgParSigDBInsufficient)
 
-		failed, step, msg = analyseDutyFailed(attDuty, events, false)
+		failed, step, msg, err = analyseDutyFailed(attDuty, events, false)
+		require.NoError(t, err)
 		require.True(t, failed)
-		require.Equal(t, step, parSigDBThreshold)
+		require.Equal(t, step, parSigDBExternal)
 		require.Equal(t, msg, msgParSigDBInconsistent)
 
 		events[syncMsgDuty] = events[attDuty]
-		failed, step, msg = analyseDutyFailed(syncMsgDuty, events, false)
+		failed, step, msg, err = analyseDutyFailed(syncMsgDuty, events, false)
+		require.NoError(t, err)
 		require.True(t, failed)
-		require.Equal(t, step, parSigDBThreshold)
+		require.Equal(t, step, parSigDBExternal)
 		require.Equal(t, msg, msgParSigDBInconsistentSync)
 	})
 
 	t.Run("FailedAtFetcherAsRandaoFailed", func(t *testing.T) {
-		// Randao failed at parSigEx
+		// Randao failed at parSigExReceive/parSigDBExternal
+		expectedErr := errors.New("failed to query randao")
 		events := map[core.Duty][]event{
 			proposerDuty: {
 				{
-					duty: proposerDuty,
-					step: scheduler,
+					duty:    proposerDuty,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				},
 			},
 			randaoDuty: {
@@ -223,32 +258,39 @@ func TestAnalyseDutyFailed(t *testing.T) {
 					duty: proposerDuty,
 					step: parSigDBInternal,
 				},
+				{
+					duty: proposerDuty,
+					step: parSigEx,
+				},
 			},
 		}
 
-		failed, step, msg := analyseDutyFailed(proposerDuty, events, true)
+		failed, step, msg, err := analyseDutyFailed(proposerDuty, events, true)
+		require.ErrorAs(t, err, &expectedErr)
 		require.True(t, failed)
 		require.Equal(t, step, fetcher)
-		require.Equal(t, msg, msgFetcherProposerFailedRandao)
+		require.Contains(t, msg, msgFetcherProposerNoExternalRandaos)
 
 		// Randao failed at parSigDBThreshold
 		events[randaoDuty] = append(events[randaoDuty], event{
 			duty: proposerDuty,
-			step: parSigEx,
+			step: parSigDBExternal,
 		})
 
-		failed, step, msg = analyseDutyFailed(proposerDuty, events, true)
+		failed, step, msg, err = analyseDutyFailed(proposerDuty, events, true)
+		require.ErrorAs(t, err, &expectedErr)
 		require.True(t, failed)
 		require.Equal(t, step, fetcher)
-		require.Equal(t, msg, msgFetcherProposerFewRandaos)
+		require.Contains(t, msg, msgFetcherProposerFewRandaos)
 
 		// No Randaos
 		events[randaoDuty] = nil
 
-		failed, step, msg = analyseDutyFailed(proposerDuty, events, true)
+		failed, step, msg, err = analyseDutyFailed(proposerDuty, events, true)
+		require.ErrorAs(t, err, &expectedErr)
 		require.True(t, failed)
 		require.Equal(t, step, fetcher)
-		require.Equal(t, msg, msgFetcherProposerZeroRandaos)
+		require.Contains(t, msg, msgFetcherProposerZeroRandaos)
 	})
 
 	t.Run("DutySuccess", func(t *testing.T) {
@@ -257,11 +299,12 @@ func TestAnalyseDutyFailed(t *testing.T) {
 			attDuty = core.NewAttesterDuty(int64(1))
 		)
 
-		for step := scheduler; step < sentinel; step++ {
+		for step := fetcher; step < sentinel; step++ {
 			events[attDuty] = append(events[attDuty], event{step: step})
 		}
 
-		failed, step, msg := analyseDutyFailed(attDuty, events, true)
+		failed, step, msg, err := analyseDutyFailed(attDuty, events, true)
+		require.NoError(t, err)
 		require.False(t, failed)
 		require.Equal(t, zero, step)
 		require.Empty(t, msg)
@@ -270,34 +313,22 @@ func TestAnalyseDutyFailed(t *testing.T) {
 
 func TestDutyFailedStep(t *testing.T) {
 	var events []event
-	for step := scheduler; step < sentinel; step++ {
+	for step := fetcher; step < sentinel; step++ {
 		events = append(events, event{step: step})
 	}
 
 	t.Run("DutySuccess", func(t *testing.T) {
-		failed, step := dutyFailedStep(events)
+		failed, step, err := dutyFailedStep(events)
+		require.NoError(t, err)
 		require.False(t, failed)
 		require.Equal(t, zero, step)
 	})
 
 	t.Run("EmptyEvents", func(t *testing.T) {
-		f, step := dutyFailedStep([]event{})
+		f, step, err := dutyFailedStep([]event{})
+		require.NoError(t, err)
 		require.True(t, f)
 		require.Equal(t, zero, step)
-	})
-
-	t.Run("DutyFailed", func(t *testing.T) {
-		// Remove the last step (sigAgg) from the events array.
-		events = events[:len(events)-1]
-		f, step := dutyFailedStep(events)
-		require.True(t, f)
-		require.Equal(t, step, sigAgg)
-
-		// Remove the second-last step (parsigDBThreshold) from the events array.
-		events = events[:len(events)-1]
-		f, step = dutyFailedStep(events)
-		require.True(t, f)
-		require.Equal(t, step, parSigDBThreshold)
 	})
 }
 
@@ -385,18 +416,18 @@ func TestTrackerParticipation(t *testing.T) {
 	}
 
 	// Ignore failedDutyReporter part to isolate participation only.
-	tr.failedDutyReporter = func(context.Context, core.Duty, bool, step, string) {}
+	tr.failedDutyReporter = func(context.Context, core.Duty, bool, step, string, error) {}
 
 	go func() {
 		for _, td := range testData {
-			require.NoError(t, tr.SchedulerEvent(ctx, td.duty, td.defSet))
-			require.NoError(t, tr.ParSigDBInternalEvent(ctx, td.duty, td.parSignedDataSet))
+			tr.FetcherFetched(td.duty, td.defSet, nil)
+			tr.ParSigDBStoredInternal(td.duty, td.parSignedDataSet, nil)
 			for _, data := range psigDataPerDutyPerPeer[td.duty] {
-				require.NoError(t, tr.ParSigExEvent(ctx, td.duty, data))
+				tr.ParSigDBStoredExternal(td.duty, data, nil)
 			}
 			for _, pk := range pubkeys {
-				require.NoError(t, tr.ParSigDBThresholdEvent(ctx, td.duty, pk, nil))
-				require.NoError(t, tr.SigAggEvent(ctx, td.duty, pk, nil))
+				tr.SigAggAggregated(td.duty, pk, nil, nil)
+				tr.BroadcasterBroadcast(td.duty, pk, nil, nil)
 			}
 
 			// Explicitly mark the current duty as deadlined.
@@ -444,7 +475,7 @@ func TestUnexpectedParticipation(t *testing.T) {
 			}
 
 			go func(duty core.Duty) {
-				require.NoError(t, tr.ParSigExEvent(ctx, duty, core.ParSignedDataSet{pubkey: data}))
+				tr.ParSigDBStoredExternal(duty, core.ParSignedDataSet{pubkey: data}, nil)
 				analyser.deadlineChan <- duty
 				deleter.deadlineChan <- duty
 			}(d)
@@ -489,8 +520,8 @@ func TestDutyRandaoUnexpected(t *testing.T) {
 	}
 
 	go func() {
-		require.NoError(t, tr.SchedulerEvent(ctx, dutyProposer, core.DutyDefinitionSet{pubkey: core.NewProposerDefinition(testutil.RandomProposerDuty(t))}))
-		require.NoError(t, tr.ParSigExEvent(ctx, dutyRandao, core.ParSignedDataSet{pubkey: data}))
+		tr.FetcherFetched(dutyProposer, core.DutyDefinitionSet{pubkey: core.NewProposerDefinition(testutil.RandomProposerDuty(t))}, errors.New("failed to query randao"))
+		tr.ParSigDBStoredExternal(dutyRandao, core.ParSignedDataSet{pubkey: data}, nil)
 
 		analyser.deadlineChan <- dutyProposer
 		// Trim Proposer events before Randao deadline
@@ -536,8 +567,8 @@ func TestDutyRandaoExpected(t *testing.T) {
 	}
 
 	go func() {
-		require.NoError(t, tr.SchedulerEvent(ctx, dutyProposer, core.DutyDefinitionSet{pubkey: core.NewProposerDefinition(testutil.RandomProposerDuty(t))}))
-		require.NoError(t, tr.ParSigExEvent(ctx, dutyRandao, core.ParSignedDataSet{pubkey: data}))
+		tr.FetcherFetched(dutyProposer, core.DutyDefinitionSet{pubkey: core.NewProposerDefinition(testutil.RandomProposerDuty(t))}, errors.New("failed to query randao"))
+		tr.ParSigDBStoredExternal(dutyRandao, core.ParSignedDataSet{pubkey: data}, nil)
 
 		analyser.deadlineChan <- dutyProposer
 		analyser.deadlineChan <- dutyRandao
@@ -645,14 +676,16 @@ func TestFromSlot(t *testing.T) {
 		close(done)
 	}()
 
-	require.NoError(t, tr.SigAggEvent(ctx, core.NewAggregatorDuty(thisSlot), "", nil))
-	require.NoError(t, tr.ValidatorAPIEvent(ctx, core.NewProposerDuty(thisSlot), nil))
-	require.NoError(t, tr.SchedulerEvent(ctx, core.NewAggregatorDuty(thisSlot), nil))
+	tr.SigAggAggregated(core.NewAggregatorDuty(thisSlot), "", nil, nil)
+	tr.ParSigDBStoredInternal(core.NewProposerDuty(thisSlot), nil, nil)
+	tr.FetcherFetched(core.NewAggregatorDuty(thisSlot), nil, nil)
+
 	require.Empty(t, tr.events)
 	cancel()
 	<-done
 }
 
+//nolint:maintidx
 func TestAnalyseFetcherFailed(t *testing.T) {
 	const slot = 123
 	dutyAgg := core.NewAggregatorDuty(slot)
@@ -668,70 +701,122 @@ func TestAnalyseFetcherFailed(t *testing.T) {
 		events map[core.Duty][]event
 		msg    string
 		failed bool
+		err    error
 	}{
+		{
+			name: "eth2 error",
+			duty: dutyAtt,
+			events: map[core.Duty][]event{
+				dutyAtt: {event{
+					duty: dutyAtt,
+					step: fetcher,
+					stepErr: errors.Wrap(eth2http.Error{
+						Method:     http.MethodGet,
+						Endpoint:   "/eth/v1/validator/attestation_data",
+						StatusCode: 404,
+						Data:       nil,
+					}, "beacon api error"),
+				}},
+			},
+			msg:    msgFetcherBN,
+			failed: true,
+			err: errors.Wrap(eth2http.Error{
+				Method:     http.MethodGet,
+				Endpoint:   "/eth/v1/validator/attestation_data",
+				StatusCode: 404,
+				Data:       nil,
+			}, "beacon api error"),
+		},
 		{
 			name: "beacon committee selections endpoint not supported",
 			duty: dutyAgg,
 			events: map[core.Duty][]event{
 				dutyAgg: {event{
-					duty: dutyAgg,
-					step: scheduler,
+					duty:    dutyAgg,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 			},
 			msg:    msgFetcherAggregatorZeroPrepares,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
-			name: "insufficient DutyPrepareAggregator signature",
+			name: "no external DutyPrepareAggregator signatures received",
 			duty: dutyAgg,
 			events: map[core.Duty][]event{
 				dutyAgg: {event{
-					duty: dutyAgg,
-					step: scheduler,
+					duty:    dutyAgg,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 				dutyPrepAgg: {event{
 					duty: dutyPrepAgg,
 					step: parSigEx,
 				}},
 			},
+			msg:    msgFetcherAggregatorNoExternalPrepares,
+			failed: true,
+			err:    context.Canceled,
+		},
+		{
+			name: "insufficient DutyPrepareAggregator signature",
+			duty: dutyAgg,
+			events: map[core.Duty][]event{
+				dutyAgg: {event{
+					duty:    dutyAgg,
+					step:    fetcher,
+					stepErr: context.Canceled,
+				}},
+				dutyPrepAgg: {event{
+					duty: dutyPrepAgg,
+					step: parSigDBExternal,
+				}},
+			},
 			msg:    msgFetcherAggregatorFewPrepares,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
 			name: "DutyPrepareAggregator failed",
 			duty: dutyAgg,
 			events: map[core.Duty][]event{
 				dutyAgg: {event{
-					duty: dutyAgg,
-					step: scheduler,
+					duty:    dutyAgg,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 				dutyPrepAgg: {event{
 					duty: dutyPrepAgg,
-					step: parSigDBThreshold,
+					step: sigAgg,
 				}},
 			},
 			msg:    msgFetcherAggregatorFailedPrepare,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
 			name: "DutyAttester failed",
 			duty: dutyAgg,
 			events: map[core.Duty][]event{
 				dutyAgg: {event{
-					duty: dutyAgg,
-					step: scheduler,
+					duty:    dutyAgg,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 				dutyPrepAgg: {event{
 					duty: dutyPrepAgg,
-					step: sigAgg,
+					step: bcast,
 				}},
 				dutyAtt: {event{
-					duty: dutyAtt,
-					step: fetcher,
+					duty:    dutyAtt,
+					step:    fetcher,
+					stepErr: errors.New("some error"),
 				}},
 			},
 			msg:    msgFetcherAggregatorNoAttData,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
 			name: "no aggregator found",
@@ -739,83 +824,110 @@ func TestAnalyseFetcherFailed(t *testing.T) {
 			events: map[core.Duty][]event{
 				dutyAgg: {event{
 					duty: dutyAgg,
-					step: scheduler,
+					step: fetcher,
 				}},
 				dutyPrepAgg: {event{
 					duty: dutyPrepAgg,
-					step: sigAgg,
+					step: bcast,
 				}},
 				dutyAtt: {event{
 					duty: dutyAtt,
-					step: sigAgg,
+					step: bcast,
 				}},
 			},
 			msg:    "",
 			failed: false,
+			err:    nil,
 		},
 		{
 			name: "sync committee selections endpoint not supported",
 			duty: dutySyncCon,
 			events: map[core.Duty][]event{
 				dutySyncCon: {event{
-					duty: dutySyncCon,
-					step: scheduler,
+					duty:    dutySyncCon,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 			},
 			msg:    msgFetcherSyncContributionZeroPrepares,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
-			name: "insufficient DutyPrepareSyncContribution signatures",
+			name: "no external DutyPrepareSyncContribution signatures",
 			duty: dutySyncCon,
 			events: map[core.Duty][]event{
 				dutySyncCon: {event{
-					duty: dutySyncCon,
-					step: scheduler,
+					duty:    dutySyncCon,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 				dutyPrepSyncCon: {event{
 					duty: dutyPrepSyncCon,
 					step: parSigEx,
 				}},
 			},
+			msg:    msgFetcherSyncContributionNoExternalPrepares,
+			failed: true,
+			err:    context.Canceled,
+		},
+		{
+			name: "insufficient DutyPrepareSyncContribution signatures",
+			duty: dutySyncCon,
+			events: map[core.Duty][]event{
+				dutySyncCon: {event{
+					duty:    dutySyncCon,
+					step:    fetcher,
+					stepErr: context.Canceled,
+				}},
+				dutyPrepSyncCon: {event{
+					duty: dutyPrepSyncCon,
+					step: parSigDBExternal,
+				}},
+			},
 			msg:    msgFetcherSyncContributionFewPrepares,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
 			name: "DutyPrepareSyncContribution failed",
 			duty: dutySyncCon,
 			events: map[core.Duty][]event{
 				dutySyncCon: {event{
-					duty: dutySyncCon,
-					step: scheduler,
+					duty:    dutySyncCon,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 				dutyPrepSyncCon: {event{
 					duty: dutyPrepSyncCon,
-					step: parSigDBThreshold,
+					step: sigAgg,
 				}},
 			},
 			msg:    msgFetcherSyncContributionFailedPrepare,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
 			name: "DutySyncMessage failed",
 			duty: dutySyncCon,
 			events: map[core.Duty][]event{
 				dutySyncCon: {event{
-					duty: dutySyncCon,
-					step: scheduler,
+					duty:    dutySyncCon,
+					step:    fetcher,
+					stepErr: context.Canceled,
 				}},
 				dutyPrepSyncCon: {event{
 					duty: dutyPrepSyncCon,
-					step: sigAgg,
+					step: bcast,
 				}},
 				dutySyncMsg: {event{
 					duty: dutySyncMsg,
-					step: parSigDBInternal,
+					step: parSigEx,
 				}},
 			},
 			msg:    msgFetcherSyncContributionNoSyncMsg,
 			failed: true,
+			err:    context.Canceled,
 		},
 		{
 			name: "no sync committee aggregators found",
@@ -823,27 +935,49 @@ func TestAnalyseFetcherFailed(t *testing.T) {
 			events: map[core.Duty][]event{
 				dutySyncCon: {event{
 					duty: dutySyncCon,
-					step: scheduler,
+					step: fetcher,
 				}},
 				dutyPrepSyncCon: {event{
 					duty: dutyPrepSyncCon,
-					step: sigAgg,
+					step: bcast,
 				}},
 				dutySyncMsg: {event{
 					duty: dutySyncMsg,
-					step: sigAgg,
+					step: bcast,
 				}},
 			},
 			msg:    "",
 			failed: false,
+			err:    nil,
+		},
+		{
+			name: "unexpected error",
+			duty: dutyAtt,
+			events: map[core.Duty][]event{
+				dutyAtt: {
+					event{
+						duty:    dutyAtt,
+						step:    fetcher,
+						stepErr: errors.New("unexpected error"),
+					},
+				},
+			},
+			msg:    msgFetcherError,
+			failed: true,
+			err:    errors.New("unexpected error"),
 		},
 	}
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			failed, step, msg := analyseDutyFailed(test.duty, test.events, true)
-			require.Equal(t, failed, test.failed)
-			require.Equal(t, test.msg, msg)
+			failed, step, msg, err := analyseDutyFailed(test.duty, test.events, true)
+			if test.err == nil {
+				require.NoError(t, err)
+			} else {
+				require.ErrorAs(t, err, &test.err)
+			}
+			require.Equal(t, test.failed, failed)
+			require.Contains(t, msg, test.msg)
 			require.Equal(t, fetcher, step)
 		})
 	}
@@ -872,7 +1006,7 @@ func TestIsParSigEventExpected(t *testing.T) {
 			name: "DutyRandao expected",
 			duty: core.NewRandaoDuty(slot),
 			events: map[core.Duty][]event{
-				core.NewProposerDuty(slot): {event{step: scheduler, pubkey: pubkey}},
+				core.NewProposerDuty(slot): {event{step: fetcher, pubkey: pubkey}},
 			},
 			out: true,
 		},
@@ -885,7 +1019,7 @@ func TestIsParSigEventExpected(t *testing.T) {
 			name: "DutyPrepareAggregator expected",
 			duty: core.NewPrepareAggregatorDuty(slot),
 			events: map[core.Duty][]event{
-				core.NewAttesterDuty(slot): {event{step: scheduler, pubkey: pubkey}},
+				core.NewAttesterDuty(slot): {event{step: fetcher, pubkey: pubkey}},
 			},
 			out: true,
 		},
@@ -898,7 +1032,7 @@ func TestIsParSigEventExpected(t *testing.T) {
 			name: "DutyPrepareSyncContribution expected",
 			duty: core.NewPrepareSyncContributionDuty(slot),
 			events: map[core.Duty][]event{
-				core.NewSyncContributionDuty(slot): {event{step: scheduler, pubkey: pubkey}},
+				core.NewSyncContributionDuty(slot): {event{step: fetcher, pubkey: pubkey}},
 			},
 			out: true,
 		},
@@ -916,7 +1050,7 @@ func TestIsParSigEventExpected(t *testing.T) {
 			name: "DutySyncMessage expected",
 			duty: core.NewSyncMessageDuty(slot),
 			events: map[core.Duty][]event{
-				core.NewSyncContributionDuty(slot): {event{step: scheduler, pubkey: pubkey}},
+				core.NewSyncContributionDuty(slot): {event{step: fetcher, pubkey: pubkey}},
 			},
 			out: true,
 		},
@@ -930,7 +1064,7 @@ func TestIsParSigEventExpected(t *testing.T) {
 }
 
 func TestAnalyseParSigs(t *testing.T) {
-	require.Empty(t, analyseParSigs(context.Background(), nil))
+	require.Empty(t, extractParSigs(context.Background(), nil))
 
 	var events []event
 
@@ -959,7 +1093,7 @@ func TestAnalyseParSigs(t *testing.T) {
 		makeEvents(n, pubkey)
 	}
 
-	allParSigMsgs := analyseParSigs(context.Background(), events)
+	allParSigMsgs := extractParSigs(context.Background(), events)
 
 	lengths := make(map[int]string)
 	for pubkey, parSigMsgs := range allParSigMsgs {
@@ -969,4 +1103,76 @@ func TestAnalyseParSigs(t *testing.T) {
 	}
 
 	require.Equal(t, expect, lengths)
+}
+
+func TestDutyFailedMultipleEvents(t *testing.T) {
+	testErr := errors.New("test error")
+	var events []event
+	for step := fetcher; step < sentinel; step++ {
+		for i := 0; i < 5; i++ {
+			events = append(events, event{step: step, stepErr: testErr})
+		}
+	}
+
+	// Failed at last step.
+	failed, step, err := dutyFailedStep(events)
+	require.True(t, failed)
+	require.Equal(t, bcast, step)
+	require.ErrorIs(t, err, testErr)
+
+	// No Failure.
+	for step := fetcher; step < sentinel; step++ {
+		events = append(events, event{step: step})
+	}
+	failed, step, err = dutyFailedStep(events)
+	require.False(t, failed)
+	require.Equal(t, zero, step)
+	require.NoError(t, err)
+}
+
+func TestUnexpectedFailures(t *testing.T) {
+	duty := core.NewAttesterDuty(123)
+	tests := []struct {
+		name    string
+		stepErr error
+		step    step
+		events  map[core.Duty][]event
+	}{
+		{
+			name:    "consensus failed with nil error",
+			stepErr: nil,
+			step:    consensus,
+			events: map[core.Duty][]event{
+				duty: {{duty: duty, step: consensus}},
+			},
+		},
+		{
+			name:    "parsigex broadcast failed with error",
+			stepErr: errors.New("parsigex broadcast err"),
+			step:    parSigEx,
+			events: map[core.Duty][]event{
+				duty: {{duty: duty, step: parSigEx, stepErr: errors.New("parsigex broadcast err")}},
+			},
+		},
+		{
+			name:    "consensus failed with nil error",
+			stepErr: nil,
+			step:    sigAgg,
+			events: map[core.Duty][]event{
+				duty: {{duty: duty, step: sigAgg}},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		failed, step, msg, err := analyseDutyFailed(duty, test.events, false)
+		require.True(t, failed)
+		require.Equal(t, step, test.step)
+		require.Equal(t, fmt.Sprintf("unexpected failure for %s duty at %s step", duty, step), msg)
+		if test.stepErr == nil {
+			require.NoError(t, err)
+		} else {
+			require.ErrorAs(t, err, &test.stepErr)
+		}
+	}
 }
